fix: avoid signed integer UB in metricInfo and metrics counter
- cb_metricInfo: negating INT_MIN is undefined behavior. Use unsigned
  arithmetic to compute the absolute value safely.
- metrics_id_counter: change from int to unsigned int so that overflow
  wraps predictably instead of being UB.

+++ r-pkg/src/callbacks.c
@@ -216,7 +216,7 @@ static cJSON *metrics_gc_cjson(const pGEcontext gc) {
     return g;
 }
 
-static int metrics_id_counter = 0;
+static unsigned int metrics_id_counter = 0;
 
 /* --- Simple metrics cache --- */
 #define MCACHE_SIZE 512
@@ -369,9 +369,9 @@ static void cb_metricInfo(int c, const pGEcontext gc,
         return;
     }
 
-    int cc = c < 0 ? -c : c;
+    unsigned int cc = c < 0 ? -(unsigned int)c : (unsigned int)c;
     char key[16];
-    snprintf(key, sizeof(key), "c%d", cc);
+    snprintf(key, sizeof(key), "c%u", cc);
     unsigned int h = mcache_hash(key, (int)strlen(key), gc);
     mcache_entry_t *cached = mcache_lookup(h);
     if (cached) {
